requirements.txt
flask
flask_sqlalchemy
flask_cors
langchain
openai
python-dotenv
